fix(extensions): rename Extension list
